requirements.txt
python-telegram-bot>=22.0,<23.0
google-api-python-client>=2.0,<3.0
google-auth>=2.0,<3.0
google-auth-oauthlib>=1.0,<2.0
